Separate AND and MERGE functionality
